Folder + Gitignore Refactoring
