+++ .gitignore
@@ -1,43 +1,14 @@
 # Default ignored files
 /shelf/
 /workspace.xml
-/.idea
+.idea/
+.vscode/
 
-# OSX
-#
-.DS_Store
-
-# Xcode
-#
 build/
-*.pbxuser
-!default.pbxuser
-*.mode1v3
-!default.mode1v3
-*.mode2v3
-!default.mode2v3
-*.perspectivev3
-!default.perspectivev3
-xcuserdata
-*.xccheckout
-*.moved-aside
-DerivedData
-*.hmap
-*.ipa
-*.xcuserstate
+.gradle/
+SoftServeCarMobile/.gradle/
 
-# Android/IntelliJ
-#
-.idea
-.gradle
-local.properties
-*.iml
-
-# node.js
-#
-node_modules/
-npm-debug.log
-yarn-error.log
+SoftServeCarMobile/node_modules/
 
 # BUCK
 buck-out/
@@ -45,20 +16,6 @@ buck-out/
 *.keystore
 !debug.keystore
 
-# fastlane
-#
-# It is recommended to not store the screenshots in the git repo. Instead, use fastlane to re-generate the
-# screenshots whenever they are needed.
-# For more information about the recommended setup visit:
-# https://docs.fastlane.tools/best-practices/source-control/
 
-*/fastlane/report.xml
-*/fastlane/Preview.html
-*/fastlane/screenshots
 
-# Bundle artifact
-*.jsbundle
 
-# CocoaPods
-/ios/Pods/
-SoftServeCarMobile/SoftServeCarMobile/SoftServeCarMobile/.buckconfig
