--- SoftServeCarMobile/.gitattributes
@@ -1,1 +1,0 @@
-*.pbxproj -text
