+++ SoftServeCarMobile/api-service/preferences-service/PreferencesService.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import APIService from '../APIService';
+import {UserPreferences} from '../../models/UserPreferences';
+import { injectable } from 'tsyringe';
+
+@injectable()
+
+class PreferencesService {
+    constructor(private apiService: APIService) { }
+    routePrefix: string = 'UserPreferences';
+
+    getUserPreferences(userId: number) {
+        return this.apiService.get<UserPreferences>(this.routePrefix + '/'+ userId);
+    }
+
+    updateUserPreferences(preferences: UserPreferences) {
+        return this.apiService.put<UserPreferences>(this.routePrefix, { data: preferences });
+    }
+}
+export default PreferencesService;
